Разделение загрузки и валидации по модулям
