+++ js/main.js
@@ -1,7 +1,8 @@
 import {generatePhotos} from './data.js';
 import {renderPictures} from './picture.js';
 import {setPicturesListener} from './full-size-picture.js';
-import './upload-file.js';
+import './upload.js';
+// import './validation.js';
 
 const DATA_AMOUNT = 25;
 const photos = generatePhotos(DATA_AMOUNT);
